requirements.txt
astroid==3.3.9
black==25.1.0
click==8.1.8
dill==0.3.9
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pylint==3.3.6
PyYAML==6.0.2
tomlkit==0.13.2
ttkbootstrap==1.10.1
